\name{countData}
\alias{countData}
\title{countData}
\value{
A dataframe.
}
\description{
  Simulated raw counts to use as input for DGE and plotPCA functions. metadata is available.
}
\usage{
  data(countData)
}
\format{
  dataframe
}
\details{
  A dataframe with 2500 rows and 4 columns (sample names).
}

