% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_SPID.R
\name{enrichr_download}
\alias{enrichr_download}
\title{Download data from enrichr in the form of a named list - function from hypeR}
\usage{
enrichr_download(
  genesets,
  db = c("Enrichr", "YeastEnrichr", "FlyEnrichr", "WormEnrichr", "FishEnrichr")
)
}
\arguments{
\item{genesets}{A name corresponding to available genesets}

\item{db}{A species}
}
\value{
A list of genesets
}
\description{
Download data from enrichr in the form of a named list - function from hypeR
}
\examples{
ATLAS <- enrichr_download("Human_Gene_Atlas")

}
