% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Check if EnrichR is online}
\usage{
getEnrichr(url, ...)
}
\arguments{
\item{url}{(Required). URL address to check}

\item{...}{(Optional). Additional parameters to pass to GET}
}
\value{
Invisible response object from GET request if successful, or NULL on failure
}
\description{
Helper function
}
\details{
Helper function for HTTP methods GET
}
