\name{ind_deg}
\alias{ind_deg}
\title{ind_deg}
\value{
A list of dataframes.
}
\description{
  Input list for metaRNAseq function (ind_deg) made of 2 different small Deseq2
   output samples dataframes for testing purposes: DEG1_df and DEG2_df.
}
\usage{
  data(ind_deg)
}
\format{
  dataframe list
}
\details{
  Each dataframe has 2 rows and 9 columns.
}

