% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{listEnrichrDbs}
\alias{listEnrichrDbs}
\title{Look up available databases on Enrichr}
\usage{
listEnrichrDbs()
}
\value{
A data.frame of available Enrichr databases
}
\description{
Look up available databases on Enrichr
}
\details{
Look up available databases on Enrichr
}
\examples{
dbs <- listEnrichrDbs()
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}
}
