\name{metadata}
\alias{metadata}
\title{metadata}
\value{
A dataframe.
}
\description{
  Metadata associated with countData for testing purposes (functions DGE, plotPCA).
}
\usage{
  data(metadata)
}
\format{
  dataframe.
}
\details{
  A dataframe with 4 rows (sample names) and 3 columns (samplesID, condition A and B, therapy T1 and T2).
}

