% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prova.R
\name{prova}
\alias{prova}
\title{Prova import package function}
\usage{
prova(x, mean)
}
\arguments{
\item{x}{a vector or a data.frame.}

\item{mean}{"geo" for geometric mean, "har" for harmonic mean.}
}
\value{
The geometric or harmonic mean of \code{x}.
}
\description{
Prova import package function
}
\examples{
gm = prova(trees$Volume, mean = "geo")
hm = prova(trees$Volume, mean = "har")
}
