% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum5.R
\name{Sum5}
\alias{Sum5}
\title{Add 5 to any number}
\usage{
Sum5(input_num)
}
\arguments{
\item{input_num}{A number.}
}
\value{
The sum of \code{input_num} and 5.
}
\description{
Add 5 to any number
}
\examples{
Sum5(6)

}
