% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fisher_invnorm.R
\name{combine_fisher_invnorm}
\alias{combine_fisher_invnorm}
\title{combine_fisher_invnorm}
\usage{
combine_fisher_invnorm(
  ind_deg,
  invnorm,
  fishercomb,
  adjpval = 0.05,
  output_tsv = T,
  output_filename = "combine_fisher_invnorm.tsv"
)
}
\arguments{
\item{ind_deg}{List of indipendent DEG dataframes with p-values to be combined.}

\item{invnorm}{inverse normal p-value combination technique dataframe (output of metaRNAseq)}

\item{fishercomb}{Fisher p-value combination technique dataframe (output of metaRNAseq)}

\item{adjpval}{threshold to represent as binary the Meta-Analysis output adjpval.}
}
\value{
A dataframe with \code{DEindices} and \code{DEname} of DEG at the chosen Benjamini Hochberg threshold, and
\code{TestStatistic}, \code{rawpval}, \code{adjpval}, \code{binaryadjpval} vectors for differential expression in the meta-analysis.
}
\description{
Combine Meta-Analysis results with individual DE tables
}
\section{Warning}{

Bla bla bla
}

\examples{
ind_deg = list(DEG1_df, DEG2_df, DEG3_df)
names(ind_deg) = c("DEG1_df", "DEG2_df", "DEG3_df")
comb_pval_df = combine_fisher_invnorm(ind_deg, invnorm, fishercomb, adjpval = 0.05, output_tsv = T, output_filename = "combine_fisher_invnorm.tsv")
}
\seealso{
\code{\link{hello}} for DGE analysis, and \code{\link{hello}} for metaRNASeq package info

Other aggregate functions: 
\code{\link{Annotate_SPID}()},
\code{\link{DGE}()},
\code{\link{Enrichment_barplot}()},
\code{\link{Enrichment}()},
\code{\link{Gene2SProtein}()},
\code{\link{SVenn}()},
\code{\link{TCGA_download}()},
\code{\link{metaRNAseq}()}
}
\concept{aggregate functions}
