% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE.R
\name{DGE}
\alias{DGE}
\title{DGE function}
\usage{
DGE(
  expression,
  metadata,
  Nreplica,
  design = "~condition",
  condition = "condition",
  TEST,
  CTRL,
  alpha = 0.05,
  exp_filt_ctrl = 0,
  FC_filt = 0,
  output_tsv = FALSE,
  output_filename = "DEGs.tsv"
)
}
\arguments{
\item{expression}{Dataframe with counts}

\item{metadata}{Dataframe with sample metadata}

\item{Nreplica}{Double. Minimum number of replicates in each group}

\item{design}{Design formula for DGE}

\item{condition}{Column of the metadata ti use for DGE results}

\item{TEST}{Character. sample name in metadata}

\item{CTRL}{Character. sample name in metadata}

\item{alpha}{Double. the significance cutoff used for optimizing the independent filtering (by default 0.1).
If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.}

\item{FC_filt}{Dataframe with counts}

\item{output_tsv}{Logical. If \code{TRUE}, outputs a tsv file with the results. By default, FALSE.}

\item{output_filename}{Name of the tsv output file. Default is DEGs.tsv.}
}
\value{
A dataframe with \code{DEGs}
}
\description{
Perform Differential Gene Expression Analysis of RNA-Seq Data
}
\section{Warning}{

Bla bla bla
}

\examples{
DEG1_df = DGE(expression, metadata, TEST = "T", CTRL = "N", Nreplica = 45, alpha = 0.05, output_tsv = T, output_filename = "DEG1.tsv")
}
\seealso{
\code{\link{hello}} for counts data and metadata download, and \code{\link{hello}} for Gene2SProtein analysis

Other aggregate functions: 
\code{\link{Annotate_SPID}()},
\code{\link{Enrichment_barplot}()},
\code{\link{Enrichment}()},
\code{\link{Gene2SProtein}()},
\code{\link{TCGA_download}()},
\code{\link{combine_fisher_invnorm}()},
\code{\link{metaRNAseq}()}
}
\concept{aggregate functions}
