% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Enrichment function}
\usage{
Enrichment(
  dfList,
  enrich.databases = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021", "KEGG_2021_Human", "MSigDB_Hallmark_2020",
    "WikiPathways_2016", "BioCarta_2016", "Jensen_TISSUES", "Jensen_COMPARTMENTS",
    "Jensen_DISEASES"),
  p_adj = 0.05,
  logFC = 1
)
}
\arguments{
\item{dfList}{Dataframes list}

\item{enrich.databases}{Vector of EnrichR databases to consult}

\item{p_adj}{Double. Adjusted pvalue threshold for the enrichment}

\item{avglogFC}{Double. Fold change threshold for the enrichment}
}
\value{
A list of enrichment tables for un and downregulated genes in the different enrichrdatabases
}
\description{
Perform enrichment Analysis of RNA-Seq Data
}
\section{Warning}{

Bla bla bla
}

\examples{
df1 <- data.frame (Geneid  = c("Mest", "Cdk1", "Pclaf", "Birc5"),
                  baseMean = c("13490.22", "10490.21", "8888.33", "750.33"),
                   log2FoldChange = c("5.78", "6.78", "7.78", "8.78"), 
                   padj = c("2.28-143", "2.18-115", "2.18-45", "0.006"))
 df2 <- data.frame (Geneid  = c("Mest", "Cdk1", "Pclaf", "Birc5"),
                   baseMean = c("13490.22", "10490.21", "8888.33", "750.33"),
                   log2FoldChange = c("5.78", "6.78", "7.78", "8.78"),
                   padj = c("2.28-143", "2.18-115", "2.18-45", "0.006"))
dfList = list(df1, df2)
names(dfList)<-c("df1", "df2")
test = Enrichment(dfList)
}
\seealso{
\code{\link{hello}} for counts data and metadata download, and \code{\link{hello}} for Gene2SProtein analysis

Other aggregate functions: 
\code{\link{DGE}()},
\code{\link{Enrichment_barplot}()},
\code{\link{Gene2SProtein}()},
\code{\link{TCGA_download}()},
\code{\link{combine_fisher_invnorm}()},
\code{\link{metaRNAseq}()}
}
\concept{aggregate functions}
