% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_barplot.R
\name{Enrichment_barplot}
\alias{Enrichment_barplot}
\title{Enrichment_barplot}
\usage{
Enrichment_barplot(
  dfList,
  enrich.databases = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021"),
  p_adj = 0.05,
  num_term = 10,
  cond = "UP"
)
}
\arguments{
\item{dfList}{A list of enrichment tables for up and down-regulated genes in the different enrichrdatabases. Output of Enrichment.R function.}

\item{enrich.databases}{Vector of EnrichR databases to consider}

\item{p_adj}{Double. Minimum Adjusted pvalue threshold for the enrichment}

\item{num_term}{Double. Number of up-regulated and dw-regulated terms to represent}

\item{cond}{String. Title of the plot.}
}
\description{
Barplot representing the top up-regulated or down-regulated significant pathways
}
\section{Warning}{

Bla bla bla
}

\examples{
Enrichment_barplot(dfList, enrich.databases = c("GO_Biological_Process_2018","GO_Cellular_Component_2018"), p_adj = 0.01, num_term = 5, cond = "UP")
}
\seealso{
\code{\link{hello}} for counts data and metadata download, and \code{\link{hello}} for Gene2SProtein analysis

Other aggregate functions: 
\code{\link{DGE}()},
\code{\link{Enrichment}()},
\code{\link{Gene2SProtein}()},
\code{\link{TCGA_download}()},
\code{\link{combine_fisher_invnorm}()},
\code{\link{metaRNAseq}()}
}
\concept{aggregate functions}
