% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_barplot.R
\name{Enrichment_barplot}
\alias{Enrichment_barplot}
\title{Enrichment_barplot}
\usage{
Enrichment_barplot(
  Enrich,
  enrich.databases = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021"),
  p_adj = 0.05,
  num_term = 10,
  cond = "UP"
)
}
\arguments{
\item{Enrich}{A list of enrichment tables for up and down-regulated genes in the different enrichR databases.
Output of Enrichment.R function for one DGE experiment.}

\item{enrich.databases}{Vector of EnrichR databases to consider. These databases must be present in the Enrich list.}

\item{p_adj}{Double. Minimum Adjusted pvalue threshold for the enrichment}

\item{num_term}{Double. Number of up-regulated and dw-regulated terms to represent}

\item{cond}{String. Title of the plot.}
}
\description{
Barplot representing the top up-regulated or down-regulated significant pathways
}
\examples{
dbs = c("GO_Biological_Process_2021","WikiPathways_2016", "MSigDB_Hallmark_2020")
dfList <- list()
if (requireNamespace("enrichR", quietly = TRUE)) {
    up_genes = c("RUNX1", "DLK1", "TOP2A", "EPCAM", "GATA1", "KDR")
    dfList[["fdr_up"]] <- enrichR::enrichr(up_genes, dbs)
    dw_genes = c("CD275", "COL1A1", "COL1A2","LUM", "SOX9")
    dfList[["fdr_down"]] <- enrichR::enrichr(dw_genes, dbs)
    # Plot for upregulated gene
    Enrichment_barplot(dfList,
                       enrich.databases = dbs,
                       p_adj = 0.01, num_term = 5, cond = "UP")
    # Plot for downregulated gene
    Enrichment_barplot(dfList,
                       enrich.databases = dbs,
                       p_adj = 0.01, num_term = 5, cond = "DOWN")
} else {
    print("example require enrichR package")
}
}
\seealso{
Other functional-annotation functions: 
\code{\link{Annotate_SPID}()},
\code{\link{Enrichment}()}

Other plot functions: 
\code{\link{SVenn}()},
\code{\link{Splot}()},
\code{\link{plotPCA}()}
}
\concept{functional-annotation functions}
\concept{plot functions}
