% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_barplot.R
\name{Enrichment_barplot}
\alias{Enrichment_barplot}
\title{Enrichment_barplot}
\usage{
Enrichment_barplot(
  Enrich,
  enrich.databases = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021"),
  p_adj = 0.05,
  num_term = 10,
  cond = "UP"
)
}
\arguments{
\item{Enrich}{A list of enrichment tables for up and down-regulated genes in the different enrichR databases.
Output of Enrichment.R function for one DGE experiment.}

\item{enrich.databases}{Vector of EnrichR databases to consider. These databases must be present in the Enrich list.}

\item{p_adj}{Double. Minimum Adjusted pvalue threshold for the enrichment}

\item{num_term}{Double. Number of up-regulated and dw-regulated terms to represent}

\item{cond}{String. Title of the plot.}
}
\description{
Barplot representing the top up-regulated or down-regulated significant pathways
}
\examples{
Enrichment_barplot(dfList,
                   enrich.databases = c("GO_Biological_Process_2018","GO_Cellular_Component_2018"),
                   p_adj = 0.01, num_term = 5, cond = "UP")
}
\seealso{
Other functional-annotation functions: 
\code{\link{Annotate_SPID}()},
\code{\link{Enrichment}()}

Other plot functions: 
\code{\link{SVenn}()},
\code{\link{Splot}()},
\code{\link{plotPCA}()}
}
\concept{functional-annotation functions}
\concept{plot functions}
