% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{enrichr}
\alias{enrichr}
\title{Gene enrichment using Enrichr}
\usage{
enrichr(genes, databases)
}
\arguments{
\item{genes}{(Required). Character vector of Entrez gene symbols as input. A data.frame
of gene symbols in first column is also acceptable, optionally a score denoting the
degree of membership between 0 and 1 in the second column.}

\item{databases}{(Required). Character vector of databases to search.
See https://maayanlab.cloud/Enrichr/ for available databases.}

\item{background}{(Optional). Character vector of Entrez gene symbols to be used as
background. A data.frame of gene symbols in first column is also acceptable.
Default is \code{"NULL"}. Enrichment analysis with background genes is only available
on the main site (Enrichr). Also, it is using a different API service (Speedrichr),
hence it is a little slower to complete and return the results.}

\item{include_overlap}{(Optional). Download database in GMT format to include 'Overlap'
in the resulting data.frame when analysing with a background. Default is \code{"FALSE"}.}
}
\value{
Returns a list of data.frame of enrichment terms, p-values, ...
}
\description{
Gene enrichment using Enrichr
}
\details{
Gene enrichment using Enrichr, slighthly modified by Aurora Maurizio.
}
\examples{
df1 <- data.frame(GeneID  = c("MEST", "CDK1", "PCLAF", "BIRC5"),
                  baseMean = c(13490.22, 10490.23, 8888.33, 750.33),
                  log2FoldChange = c(5.78, 6.76, -7.78, -8.78),
                  padj = c(2.28e-143, 2.18e-115, 2.18e-45, 0.006),
                  row.names = c("MEST", "CDK1", "PCLAF", "BIRC5"))
df2 <- data.frame(GeneID  = c("MEST", "CDK1", "PCLAF", "BIRC5"),
                  baseMean = c(13490.22, 10490.23, 8888.33, 750.33),
                  log2FoldChange = c(5.78, 6.76, -7.78, -8.78),
                  padj = c(2.28e-143, 2.18e-115, 2.18e-45, 0.006),
                  row.names = c("MEST", "CDK1", "PCLAF", "BIRC5"))
dfList <- list(df1 = df1, df2 = df2)
dbs <- c("GO_Molecular_Function_2023", "GO_Cellular_Component_2023",
         "GO_Biological_Process_2023")
enriched1 <- enrichr(input, dbs)
print(head(enriched1[[1]]))
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}
}
