% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene2SProtein.R
\name{Gene2SProtein}
\alias{Gene2SProtein}
\title{Detect Surface Proteins from a list of genes}
\usage{
Gene2SProtein(genes, output_tsv)
}
\arguments{
\item{genes}{array of gene names.}

\item{output_tsv}{output a tsv file with the results - boolean variable.}
}
\value{
A dataframe of surface proteins filtered from the \code{genes} array.
}
\description{
Detect Surface Proteins from a list of genes
}
\examples{
GeneNames = c("CIITA", "EPCAM", "DLK1", "CD24", "CDCP1", "LYVE1", "ABCD1", "VAMP1")
SurfaceProteins_df = Gene2SProtein(GeneNames, output_tsv = FALSE)
}
