% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaRNAseq.R
\name{metaRNAseq}
\alias{metaRNAseq}
\title{metaRNAseq function}
\usage{
metaRNAseq(
  ind_deg,
  test_statistic = "fishercomb",
  BHth = 0.05,
  adjpval.t = 0.05,
  nrep = NULL
)
}
\arguments{
\item{ind_deg}{List of indipendent named DEG dataframes with p-values to be combined.}

\item{test_statistic}{p-value combination technique (inverse normal or Fisher):
\code{fishercomb}, \code{invnorm}.
By default: \code{fishercomb}.}

\item{BHth}{Benjamini Hochberg threshold.}

\item{adjpval.t}{threshold to represent as binary the Meta-Analysis output adjpval.
By default, the False Discovery Rate is controlled at 5%.}

\item{nrep}{Vector of numbers of replicates used in each study to calculate the previous one-sided p-values.}
}
\value{

}
\description{
Perform Meta-Analysis of RNA-Seq Data
}
\section{Warning}{

Bla bla bla
}

\examples{
ind_deg = list(DEG1_df, DEG2_df, DEG3_df)
names(ind_deg) = c("DEG1_df", "DEG2_df", "DEG3_df")
comb_pval_df = Meta(ind_deg, test_statistic = "invnorm", BHth = 0.05, nrep = c(2,2,2))
}
\seealso{
\code{\link{hello}} for DGE analysis, and \code{\link{hello}} for metaRNASeq package info

Other aggregate functions: 
\code{\link{Annotate_SPID}()},
\code{\link{DGE}()},
\code{\link{Enrichment_barplot}()},
\code{\link{Enrichment}()},
\code{\link{Gene2SProtein}()},
\code{\link{TCGA_download}()},
\code{\link{combine_fisher_invnorm}()}
}
\concept{aggregate functions}
