% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA function}
\usage{
plotPCA(
  matrix,
  metadata,
  nTOP = 500,
  dims = c(1, 2),
  color.by = NULL,
  pt.shape = NULL,
  cols.use = NULL,
  label = FALSE,
  new.label = NULL,
  label.size = 5
)
}
\arguments{
\item{matrix}{Count matrix with gene on the row and sample ID on the column.}

\item{metadata}{Sample metadata}

\item{nTOP}{number of top genes to use for principal components,
selected by highest row variance}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{color.by}{Name of one or more metadata columns to color cells by.}

\item{pt.shape}{If NULL, all points are circles \(default\).
You can specify any metadata attribute allowing for both different colors
and different shapes on cells.}

\item{cols.use}{Vector of colors, each color corresponds to an identity class.
By default, ggplot assigns colors.}

\item{label}{Logical. If \code{TRUE} adds samples label. Default = FALSE.}

\item{new.label}{If NULL, use the sample names as in metadata.
Otherwise you can specify new labels.}

\item{label.size}{Sets size of labels}
}
\value{
PCA plot objec created by ggplot2,
which can be assigned and further customized.
}
\description{
Plot PCA highlighting one or two data features
}
\examples{

plotPCA(matrix = count_GSE133671,
        metadata = meta_GSE133671,
        color = "GSE",
        shape = "condition")
}
