% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVenn.R
\name{SVenn}
\alias{SVenn}
\title{SVenn}
\usage{
SVenn(
  S_List,
  cols.use = NULL,
  main = "Venn",
  opacity = 0.5,
  output_intersection = FALSE,
  filename = "intersection.xlsx"
)
}
\arguments{
\item{cols.use}{Vector of colors, each color corresponds to a study.
By default, ggplot assigns colors.}

\item{main}{Plot title. Default = Venn.}

\item{opacity}{Degree of opacity for the color(s) specified with cols.use (less opacity, more transparency).}

\item{output_intersection}{logical. If TRUE write an output of protein in the intersections}

\item{filename}{Name of the output file with the intersections}

\item{S_list}{A list of surface proteins detected in different studies.}
}
\value{
a
}
\description{
Venn diagram of common surface proteins overexpressed among different studies
}
\examples{
SVenn(S_list, main = "my Venn")
}
\seealso{
\code{\link{hello}} for counts data and metadata download, and \code{\link{hello}} for Gene2SProtein analysis

Other plot functions: 
\code{\link{Enrichment_barplot}()},
\code{\link{Splot}()},
\code{\link{plotPCA}()}
}
\concept{plot functions}
