% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{TCGA_download}
\alias{TCGA_download}
\title{TCGA_download function}
\usage{
TCGA_download(
  project,
  whichcounts = "unstranded",
  save.matrix = F,
  save.metadata = F
)
}
\arguments{
\item{project, }{Character. A valid project from TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{whichcounts}{Character. Counts data form to use. Choose from: unstranded, stranded_first,stranded_second. By default, unstranded.}

\item{save.matrix}{Logical. If \code{TRUE}, outputs a tsv file with the Matrix. By default, FALSE.}

\item{save.metdata}{Logical. If \code{TRUE}, outputs a tsv file with the metadata. By default, FALSE.}
}
\value{
A list containing the Matrix and the metadata.
}
\description{
Downloads count matrix data from TCGA_download
}
\section{Warning}{

Bla bla bla
}

\examples{
TCGA_download(project="TCGA-UVM", whichcounts = "unstranded", save.matrix = F, save.metadata = T)
}
\seealso{
\code{\link{hello}} for \code{\link{hello}} for DGE analysis.

Other aggregate functions: 
\code{\link{Annotate_SPID}()},
\code{\link{DGE}()},
\code{\link{Enrichment_barplot}()},
\code{\link{Enrichment}()},
\code{\link{Gene2SProtein}()},
\code{\link{combine_fisher_invnorm}()},
\code{\link{metaRNAseq}()}
}
\concept{aggregate functions}
